requirements.txt
Django
markdown
